Fix tests, remove dead code
