+++ frontend/lib/badge-url.spec.js
@@ -6,19 +6,17 @@ const resolveBadgeUrlWithLongCache = (url, baseUrl) =>
 
 describe('Badge URL functions', function() {
   test(resolveBadgeUrl, () => {
-    given('/badge/foo-bar-blue.svg', undefined).expect(
-      '/badge/foo-bar-blue.svg'
-    )
-    given('/badge/foo-bar-blue.svg', 'http://example.com').expect(
+    given('/badge/foo-bar-blue', undefined).expect('/badge/foo-bar-blue.svg')
+    given('/badge/foo-bar-blue', 'http://example.com').expect(
       'http://example.com/badge/foo-bar-blue.svg'
     )
   })
 
   test(resolveBadgeUrlWithLongCache, () => {
-    given('/badge/foo-bar-blue.svg', undefined).expect(
+    given('/badge/foo-bar-blue', undefined).expect(
       '/badge/foo-bar-blue.svg?maxAge=2592000'
     )
-    given('/badge/foo-bar-blue.svg', 'http://example.com').expect(
+    given('/badge/foo-bar-blue', 'http://example.com').expect(
       'http://example.com/badge/foo-bar-blue.svg?maxAge=2592000'
     )
   })
