--- frontend/lib/prepare-examples.js
@@ -1,42 +1,0 @@
-import escapeStringRegexp from 'escape-string-regexp'
-
-export function exampleMatchesRegex(example, regex) {
-  const { title, keywords } = example
-  const haystack = [title].concat(keywords).join(' ')
-  return regex.test(haystack)
-}
-
-export function predicateFromQuery(query) {
-  const escaped = escapeStringRegexp(query)
-  const regex = new RegExp(escaped, 'i') // Case-insensitive.
-  return ({ examples }) =>
-    examples.some(example => exampleMatchesRegex(example, regex))
-}
-
-export function mapExamples(categories, iteratee) {
-  return (
-    categories
-      .map(({ category, examples }) => ({
-        category,
-        examples: iteratee(examples),
-      }))
-      // Remove empty categories.
-      .filter(({ category, examples }) => examples.length > 0)
-  )
-}
-
-export function prepareExamples(categories, predicateProvider) {
-  let nextKey = 0
-  return mapExamples(categories, examples =>
-    examples.map(example =>
-      Object.assign(
-        {
-          shouldDisplay: () => predicateProvider()(example),
-          // Assign each example a unique ID.
-          key: nextKey++,
-        },
-        example
-      )
-    )
-  )
-}
